Externalized H2 DB configs
Externalized db path so that the DB file can be configured to be put outside the base application. By default the path is blank in which case it would remain as your default original path ../WEB-INF/classes/keybd

Moved all Connection Pool setting to the config file as well to allow for increasing the size or modifying other params related to DB Pools.